move dice icons to context and import dice from context
